move srcfiles to src directory
