+++ src/components/organisms/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Contact from '../../pages/Contact'
+import Contact from '../../../pages/contact'
 import { Logo } from '../atoms/Logo'
 
 export const Footer = () => {
