+++ src/lib/notion.ts
@@ -5,7 +5,7 @@ export const notion = new Client({
   auth: process.env.NOTION_TOKEN
 })
 
-export const getDatabase = async (databaseId) => {
+export const getDatabase = async (databaseId: string) => {
   const response = await notion.databases.query({
     database_id: databaseId
   })
