+++ src/pages/contact.tsx
@@ -1,11 +1,11 @@
 import Link from 'next/link'
 import React from 'react'
-import { Hero } from '../components/organisms/Hero'
+import { Hero } from '@/components/organisms/Hero'
 
-export default function work() {
+export default function Contact() {
   return (
     <div className="container">
-      <Hero title="Work" />
+      <Hero title="Contact" />
     </div>
   )
 }
