+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { Hero } from '../components/organisms/Hero'
+import { Hero } from '@/components/organisms/Hero'
 
 export default function Home() {
   return (
